refactor: rename MainActivity to ScreeningActivity
