+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
             android:name=".view.ticket.TicketActivity"
             android:exported="false" />
         <activity
-            android:name=".view.MainActivity"
+            android:name=".view.screening.ScreeningActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
