+++ app/src/main/res/layout/activity_screening.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ListView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/lv_main_movies"
+    android:id="@+id/lv_screening_movies"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.MainActivity"
+    tools:context=".view.screening.ScreeningActivity"
     tools:listitem="@layout/item_screening" />
